Switched back to single pass for populating binary properties on tombstone. Added log statement when a destroy operation fails to run due to having no objects to destroy. DestroyProxyService return null instead of failing if no proxy for the provided object. Reduced caching period for tests to resolve out of date objects from being returned in quick integration tests.

+++ persistence/src/main/java/edu/unc/lib/dl/persist/services/destroy/DestroyProxyService.java
@@ -27,6 +27,8 @@ import org.apache.jena.rdf.model.Resource;
 import org.fcrepo.client.FcrepoClient;
 import org.fcrepo.client.FcrepoOperationFailedException;
 import org.fcrepo.client.FcrepoResponse;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import edu.unc.lib.dl.fedora.PID;
 import edu.unc.lib.dl.fedora.ServiceException;
@@ -40,6 +42,8 @@ import edu.unc.lib.dl.sparql.SparqlQueryService;
  *
  */
 public class DestroyProxyService {
+
+    private static final Logger log = LoggerFactory.getLogger(DestroyProxyService.class);
 
     private SparqlQueryService sparqlQueryService;
     private FcrepoClient fcrepoClient;
@@ -55,6 +59,10 @@ public class DestroyProxyService {
      */
     public String destroyProxy(PID objPid) {
         ProxyInfo proxyInfo = getProxyInfo(objPid);
+        if (proxyInfo == null) {
+            log.debug("No proxy found for object {}", objPid);
+            return null;
+        }
         URI proxyUri = proxyInfo.proxyUri;
 
         try (FcrepoResponse resp = fcrepoClient.delete(proxyUri).perform()) {
